convert apollo-utilities to use jest
