+++ packages/apollo-utilities/src/__tests__/fragments.ts
@@ -1,83 +1,85 @@
 import { print } from 'graphql/language/printer';
 import gql from 'graphql-tag';
+import { disableFragmentWarnings } from 'graphql-tag';
 
-import { assert } from 'chai';
+// Turn off warnings for repeated fragment names
+disableFragmentWarnings();
 
-import { getFragmentQueryDocument } from '../src/fragments';
+import { getFragmentQueryDocument } from '../fragments';
 
 describe('getFragmentQueryDocument', () => {
   it('will throw an error if there is an operation', () => {
-    assert.throws(
-      () =>
-        getFragmentQueryDocument(
-          gql`
-            {
-              a
-              b
-              c
-            }
-          `,
-        ),
+    expect(() =>
+      getFragmentQueryDocument(
+        gql`
+          {
+            a
+            b
+            c
+          }
+        `,
+      ),
+    ).toThrowError(
       'Found a query operation. No operations are allowed when using a fragment as a query. Only fragments are allowed.',
     );
-    assert.throws(
-      () =>
-        getFragmentQueryDocument(
-          gql`
-            query {
-              a
-              b
-              c
-            }
-          `,
-        ),
+    expect(() =>
+      getFragmentQueryDocument(
+        gql`
+          query {
+            a
+            b
+            c
+          }
+        `,
+      ),
+    ).toThrowError(
       'Found a query operation. No operations are allowed when using a fragment as a query. Only fragments are allowed.',
     );
-    assert.throws(
-      () =>
-        getFragmentQueryDocument(
-          gql`
-            query Named {
-              a
-              b
-              c
-            }
-          `,
-        ),
+    expect(() =>
+      getFragmentQueryDocument(
+        gql`
+          query Named {
+            a
+            b
+            c
+          }
+        `,
+      ),
+    ).toThrowError(
       "Found a query operation named 'Named'. No operations are allowed when using a fragment as a query. Only fragments are allowed.",
     );
-    assert.throws(
-      () =>
-        getFragmentQueryDocument(
-          gql`
-            mutation Named {
-              a
-              b
-              c
-            }
-          `,
-        ),
+    expect(() =>
+      getFragmentQueryDocument(
+        gql`
+          mutation Named {
+            a
+            b
+            c
+          }
+        `,
+      ),
+    ).toThrowError(
       "Found a mutation operation named 'Named'. No operations are allowed when using a fragment as a query. " +
         'Only fragments are allowed.',
     );
-    assert.throws(
-      () =>
-        getFragmentQueryDocument(
-          gql`
-            subscription Named {
-              a
-              b
-              c
-            }
-          `,
-        ),
+    expect(() =>
+      getFragmentQueryDocument(
+        gql`
+          subscription Named {
+            a
+            b
+            c
+          }
+        `,
+      ),
+    ).toThrowError(
       "Found a subscription operation named 'Named'. No operations are allowed when using a fragment as a query. " +
         'Only fragments are allowed.',
     );
   });
 
   it('will throw an error if there is not exactly one fragment but no `fragmentName`', () => {
-    assert.throws(() => {
+    expect(() => {
       getFragmentQueryDocument(gql`
         fragment foo on Foo {
           a
@@ -91,8 +93,10 @@ describe('getFragmentQueryDocument', () => {
           f
         }
       `);
-    }, 'Found 2 fragments. `fragmentName` must be provided when there is not exactly 1 fragment.');
-    assert.throws(() => {
+    }).toThrowError(
+      'Found 2 fragments. `fragmentName` must be provided when there is not exactly 1 fragment.',
+    );
+    expect(() => {
       getFragmentQueryDocument(gql`
         fragment foo on Foo {
           a
@@ -112,16 +116,20 @@ describe('getFragmentQueryDocument', () => {
           i
         }
       `);
-    }, 'Found 3 fragments. `fragmentName` must be provided when there is not exactly 1 fragment.');
-    assert.throws(() => {
+    }).toThrowError(
+      'Found 3 fragments. `fragmentName` must be provided when there is not exactly 1 fragment.',
+    );
+    expect(() => {
       getFragmentQueryDocument(gql`
         scalar Foo
       `);
-    }, 'Found 0 fragments. `fragmentName` must be provided when there is not exactly 1 fragment.');
+    }).toThrowError(
+      'Found 0 fragments. `fragmentName` must be provided when there is not exactly 1 fragment.',
+    );
   });
 
   it('will create a query document where the single fragment is spread in the root query', () => {
-    assert.deepEqual(
+    expect(
       print(
         getFragmentQueryDocument(gql`
           fragment foo on Foo {
@@ -131,6 +139,7 @@ describe('getFragmentQueryDocument', () => {
           }
         `),
       ),
+    ).toEqual(
       print(gql`
         {
           ...foo
@@ -146,7 +155,7 @@ describe('getFragmentQueryDocument', () => {
   });
 
   it('will create a query document where the named fragment is spread in the root query', () => {
-    assert.deepEqual(
+    expect(
       print(
         getFragmentQueryDocument(
           gql`
@@ -174,6 +183,7 @@ describe('getFragmentQueryDocument', () => {
           'foo',
         ),
       ),
+    ).toEqual(
       print(gql`
         {
           ...foo
@@ -201,7 +211,7 @@ describe('getFragmentQueryDocument', () => {
         }
       `),
     );
-    assert.deepEqual(
+    expect(
       print(
         getFragmentQueryDocument(
           gql`
@@ -229,6 +239,7 @@ describe('getFragmentQueryDocument', () => {
           'bar',
         ),
       ),
+    ).toEqual(
       print(gql`
         {
           ...bar
@@ -256,7 +267,7 @@ describe('getFragmentQueryDocument', () => {
         }
       `),
     );
-    assert.deepEqual(
+    expect(
       print(
         getFragmentQueryDocument(
           gql`
@@ -284,6 +295,7 @@ describe('getFragmentQueryDocument', () => {
           'baz',
         ),
       ),
+    ).toEqual(
       print(gql`
         {
           ...baz
