--- packages/apollo-utilities/tests/tests.ts
@@ -1,14 +1,0 @@
-import { disableFragmentWarnings } from 'graphql-tag';
-
-// Turn off warnings for repeated fragment names
-disableFragmentWarnings();
-
-import './assign';
-import './cloneDeep';
-import './directives';
-import './environment';
-import './fragments';
-import './getFromAST';
-import './isEqual';
-import './transform';
-import './warnOnce';
